crm: fix bad practice of default value
In python the default value is instanciate only once.
So if the code modify the default value, it is also for all calls that comes after.
Good practice is to not use [] nor {} as default value, but use instead None or False.

bzr revid: ced-240a55be30f4d38046ca6b310967ad7742dd717e